+++ zcash_client_backend/src/data_api/wallet.rs
@@ -1033,6 +1033,9 @@ where
                 }
                 transparent_output_meta.push((to, payment.amount));
             }
+            Address::TransparentSourceOnly(_) => {
+                panic!("Transparent-source-only addresses should not occur at this stage");
+            }
         }
     }