+++ zcash_client_sqlite/src/wallet/init/migrations/ufvk_support.rs
@@ -131,7 +131,7 @@ impl<P: consensus::Parameters> RusqliteMigration for Migration<P> {
                             });
                         }
                     }
-                    Address::Transparent(_) => {
+                    Address::Transparent(_) | Address::TransparentSourceOnly(_) => {
                         return Err(WalletMigrationError::CorruptedData(
                             "Address field value decoded to a transparent address; should have been Sapling or unified.".to_string()));
                     }
@@ -265,7 +265,9 @@ impl<P: consensus::Parameters> RusqliteMigration for Migration<P> {
                     Address::Sapling(_) => {
                         Ok(pool_code(PoolType::Shielded(ShieldedProtocol::Sapling)))
                     }
-                    Address::Transparent(_) => Ok(pool_code(PoolType::Transparent)),
+                    Address::Transparent(_) | Address::TransparentSourceOnly(_) => {
+                        Ok(pool_code(PoolType::Transparent))
+                    }
                     Address::Unified(_) => Err(WalletMigrationError::CorruptedData(
                         "Unified addresses should not yet appear in the sent_notes table."
                             .to_string(),